PR feedback: use blacklist instead of whitelist approach on `ModelConfig`
Previous to this commit we were using a whitelist approach to define
the configuration options that could be override through config models,
and they were:

* `conninfo`
* `primary_conninfo`
* `streaming_conninfo`

While discussing within the team we think it would be better if it
were more flexible, so we turned that whitelist into a blacklist
instead, where we disallow overriding only the value of configuration
options which are related with paths for the Barman server, or with
hooks for the Barman server, namely:

* Path related:

    * `backup_directory`
    * `basebackups_directory`
    * `errors_directory`
    * `incoming_wals_directory`
    * `streaming_wals_directory`
    * `wals_directory`

* Hook related:

    * `post_archive_retry_script`
    * `post_archive_script`
    * `post_backup_retry_script`
    * `post_backup_script`
    * `post_delete_script`
    * `post_delete_retry_script`
    * `post_recovery_retry_script`
    * `post_recovery_script`
    * `post_wal_delete_script`
    * `post_wal_delete_retry_script`
    * `pre_archive_retry_script`
    * `pre_archive_script`
    * `pre_backup_retry_script`
    * `pre_backup_script`
    * `pre_delete_script`
    * `pre_delete_retry_script`
    * `pre_recovery_retry_script`
    * `pre_recovery_script`
    * `pre_wal_delete_script`
    * `pre_wal_delete_retry_script`

That should make it easier to introduce new settings to Barman, as
well as make config models more useful for users.

References: BAR-122.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>