Run WAL streaming checks before starting streaming
This patch runs the WAL streaming checks during
`barman.server.Server.receive_wal` before the streaming archiver is
invoked. This allows us to catch errors such as a lack of monitoring
privileges, mismatching system IDs and incorrect WAL levels.

This particular location in the code was chosen because it means the
checks are carried out when both `barman cron` and `barman receive-wal`
are executed. In the `barman cron` case the `barman receive-wal` child
process will fail leaving a trail of errors in the log but `barman cron`
itself will continue and exit with status zero. This is in line with
existing cases where `barman receive-wal` will fail for other reasons
such as issues with the replication slot.

To implement this we have to separate the WAL streaming checks into
those which should be run before WAL streaming starts (aka "preflight"
checks) and those which should be run to verify WAL streaming is working
(currently just the replication slot checks). This means we can then
call the preflight checks during `barman.server.Server.receive_wal` and
avoid failing due to the replication slot not being active.

Relates to BAR-120.