Change the minor version on the docs for rsync optionally needed
There's a requirement in the docs which is optional if you don't
use rsync mode. This stated that you needed version 3.0.4 or
better, yet this was not true as we need at least version 3.1.0
to use `--ignore-missing-args`.

This commit sets the documentation right so there's no confusion
regarding the minimum version of rsync that is required to run
barman with rsync.

This never came up because all supported OS distributions already
have rsync 3.1.2 or newer.

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>