Update instructions for barman user needed privileges
There was a block of SQL code added some time ago which helps
users that do not want to have a superuser running the backups.

These instructions provided SQL commands to GRANT those specific
privileges to the barman user.

Since version 15, two of the functions that needed to be GRANTed
have not only changed their names, but also their signatures.
There is a note regarding such change with appropriate instructions
on the new lines to use.

Given there are already two newer versions out, and Postgres 17
coming out in a some months, it makes more sense to have
instructions for the latest versions, and a note for anyone
running on older Postgres versions.

This patch amends this.

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>