Add `summarize_wal` information to backup metadata
With postgres major version >= 17, we have the possibility to do incremental
backups with pg_basebackup.

A crucial element for enabling this functionality is the WAL summarizer process.
This is a recent addition, designed to assist incremental backups by identifying
the differences in the data directory since the previous backup.

A new field was added in configuration file postgresql.conf called summarize_wal.
This field handles the state of the process (on or off).

This commit:

1. Add a new field called summarize_wal to the BackupInfo class (single backup
   metadata)
2. Add a new parameter pg_version to build_mocked_server and build_backup_manager
   methods from testing_helpers.
3. Update pieces of code accordingly where new parameters need to be handled.

References: BAR-172.

Signed-off-by: Andre Marchesini de Aguiar andre.marchesini@enterprisedb.com