Add and modify unit tests accordingly to implementations
This commit:
1. Modifies the test_walk_backups_tree unit test to also handle
the new parameter return_self.
2. Modifies the test_backup_status unit test to also handle
getting the status of incremental backups. Since the status is
propagated from the parent, this modifications also tests if
the child has the correct status.
3. Adds a unit test test__propagate_retention_status_to_children
for the new _propagate_retention_status_to_children method.
- This tests the propagation of the retention status from parent
to children backups (full to incremental backups) for the retention
policy.

References: BAR-166

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>