Refactoring broken tests from code implementations
This commit fixes unit tests which got broken because of changes
introduced to retention policies to handle incremental backups.
The previous unit tests got broken because there were two fields
added to the backup.info (PR#936, BAR-165) and also two properties
from this PR that were not being set when creating a mock for the
backup.info.
Also a new method _create_backup_info was created and the
_create_backup_metadata was refactored to give more flexibility
when creating mocks for the TestCloudBackupDelete class.

References: BAR-166

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>