Add new command wrapper for pg_combinebackup command
With the addition of incremental backups in Postgres 17,
there is a need to combine the backups to create a new
synthetic backup that can be used for recover. A new
command was added for this purpose called
pg_combinebackup.

This commit adds a new wrapper for the pg_combinebackup
client command.

References: BAR-168

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>