Modify recover method to reuse recovery_staging_path
With the addition of incremental backups for Postgres 17,
there is a need to combine the backups to create a new
synthetic backup so the user can recover from it.

It was agreed to use the existing recovery_staging_path
as the destination of the new synthetic backup when
recovering from an incremental before moving to the final
destination. Before this change, recovery_staging_path
was only mandatory and strictly used when recovering
from compressed backups.

This commit adds a check to abort the recover process
in case recovery_staging_path is not set as a parameter
in the cli or in the barman global/server configuration
files. This recovery_staging_path parameter will also be
mandatory when taking incremental backups.

References: BAR-168

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Gustavo William <gustavowilliam0980@gmail.com>