Fix issue with unit test
We noticed a unit test which has been failing. The failure occurs
because of a suffix which was not expected by the unit test:

```
by put-wal for server 'main' (SSH host: 10.0.2.2)\n
```

The unit test was not expecting the SSH host information. This commit
fixes the unit test so it won't care about this suffix.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>