Add cluster size to backup.info
After the introduction of incremental backups in Postgres 17, there
is the need to review how we display information about backups,
specially when using the `show-backup` command. Because of this,
we are adding the cluster size information to the backup.info.
This new field is intended to be used in the calculation of the
deduplication_ratio of backups and to show the estimated size of
the postgres cluster at the time the backup was taken.

This commit:
	1. Add a new field to the LocalBackupInfo class called
	cluster_size which represents the current size of the
	postgres server at the time the backups was taken.
	2. Modify `_pg_get_metadata` from BackupStrategy to
	set the new field `cluster_size` to the value of
	`PostgreSQLConnection` `current_size` property.

References: BAR-215

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>