Add new recovery executor for incremental backups
With the addition of incremental backups for postgres 17,
the user needs to be able to recover from one. To be able
to do that there is the need to combine all the incremental
backups and the full to create a new synthetic backup that
will be used to recover. The combine is hidden from the user
so he/she just needs to use the barman recover command with
an incremental backup.

This commit:
1. Add a new class SyntheticBackupInfo
	1.1 Override the get_basebackup_directory method
This was created to be able to reuse some methods from
LocalBackupInfo so it is easier to deal with the location
of the new synthetic backup. This new class has a new
parameter called base_directory which is used as the
destination for the new backup.
2. Add a new IncrementalRecoveryExecutor
	2.1 Add methods _start_message and _end_message
These are just to inform the user when the pg_combinebackup
started and when it finished.
	2.2 Add method _fetch_remote_status
This is used to retrieve information about the
pg_combinebackup client in the users machine. before
actually using the client command, there is the need to
check if it is installed and version.
	2.3 Add _retry_handler
This is just to use as the retry and warn the user of any
problem when combining. If there is a problem, the retry
will remove and recreate the destination directories.
	2.4 Add _prepare_destination and _move_destination
methods
These are used to prepare the destination directories/files
(remove if exists and create) and to move the directories/files
from source to destination (staging path to final destination).
	2.5 Add _get_backup_chain_path
This gets the list of backups that make a chain. The list is
created in the correct order starting from full backup to the
last incremental backup (the one that is being recovered from).
	2.6 Add _combine_backups method
This is the method responsible to combine the backups and
create a new synthetic backup. It uses a few of the
aforementioned methods. It combines the destination in the
recovery_staging_path.
	2.7 Override _backup_copy method
This method moves the backup to the final destination if the
recover is local. If the recover is remote, it uses the
_backup_copy from the super class.
	2.8 Override recover method
This is only used to previously combine the backup and then
use the recover from the parent class.

References: BAR-167 & BAR-168

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Gustavo William <gustavowilliam0980@gmail.com>
Co-authored-by: Barbara Leidens <barbara.leidens@enterprisedb.com>