Check that summarize_wal is enabled on incremental backups
When performing an incremental backup using postgres mode
with the new '--incremental' flag it is required that
the user has 'summarizer_wal' enabled on the postgres server.

This commit adds a new validation method which is called before
starting the backup which will perform this check and issue an error
to the user if the configurations are not met.

References: BAR-160

Signed-off-by: Gustavo William <gustavowilliam0980@gmail.com>