Verify that the Postgres version is 17 or newer on incremental backups
Performing an incremental backup using Postgres mode requires the
Postgres server's version to be 17 or newer. This PR builds from
the validation methods introduced by BAR-160 (`validate_backup_args`
and `_validate_incremental_backup_configs`), adding a check for the
Postgres server's version and issue an error if it's not compatible
with incremental backups (<17).

References: BAR-163

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>