Incremental backups tree walking
This patch adds methods to walk a tree of incremental backups.

The `walk_backups_tree` method, given a node as root, traverses the tree
in post-order identifying leaves first and nodes parent of the leaves
after. This is useful for actions like: removal of a full chain of
backups. The method is implemented using yield statements, that allow to
iterate through the chain of nodes

The `walk_to_root` method allows to reach the parent full backup of a
given backup node of an incremental backups tree. The method is
implemented using yield statements allowing to cycle from the node until
the root.

References: bar-182

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>
Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>