Add property deduplication_ratio to LocalBackupInfo
Since the introduction of incremental backups, there is the
need to calculate the deduplication_ratio also for incrementals,
not only for rsync backups.

This commit is moving the responsability to calculate this to
the LocalBackupInfo class as a property instead of repeating the
calculation in some places of the code base.

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>