Always add an explicit time zone to `recovery_target_time`
Previous to this commit there could be cases where Postgres would be
misconfigured by `barman recover`.

For example, assume you had an environment like this:

* Barman host configured with time zone UTC+2, e.g. `Europe/Berlin`;`
* Postgres cluster with default value for timezone GUC, i.e. GMT value.

If we attempted a `barman recover` execution with `--target-time` set,
but without an explicit time zone, recovery could fail:

* `barman recover` output would say that recovery target time was set
  with `+02:00`;
* However `postgresql.auto.conf` would not contain `+02:00`.

In practice either the recovery would fail because of an invalid
recovery target, or the cluster would be recovered to a point-in-time
different from the user needs. Neither of these cases are welcome.

This commit fixes that issue by reusing the `target_datetime` key from
the `recovery_info` variable. That entry contains the target time with
a time zone explicitily set, even if the user specified no time zone
through the command-line option.

Note: a warning is issued if the user specified no time zone, and Barman
assumed its host time zone.

References: BAR-185 #881.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>