Delete backup children when a parent is deleted
When a postgres full backup with incrementals is deleted
we have to make sure that all its descendants are deleted
along with it. Similarly, when a child backup is deleted
we have to make sure that its reference is removed from
is parent. To achieve this, the following changes were made:

1 - Some validations (check keep and check redundancy) were
moved from the backup manager to the server as they now will
need to happen in the very first stage of the deletion, which
happens in the server context. This becomes required once we have
to make sure that the parent itself can be deleted before proceding
with the deletion of any of its children.

2 - Split the 'delete' method of the server into two separate methods so
it gets easier to distinguish the prepare phase (where children are gathered) and
the actual delete phase (where the backup manager is called).

3 - The whole chain deletion will happen inside a server lock and every
individual backup deletion will happen inside backup lock. This ensures that
the whole process is done consistently, thus avoiding errors in the
middle which would require the user to try again.

References: BAR-181

Signed-off-by: Gustavo William <gustavowilliam0980@gmail.com>