Restrict the use of keep annotations to full backups only
An incremental backup with the keep annotation that has its
parent deleted would become an orphan backup and therefore
useless for recovery. To avoid this, we should restrict the
use of keep annotations to full backups only. This adds a
new check prior to starting the annotation process to make
sure we do not allow keeping a backup if it has a parent
backup, raising an error if so.

References: BAR-184

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Co-authored-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>