When recovering with `--no-get-wal` and `--target-lsn`, bring only the WALs up to the specified LSN
Previous to this commit, when the user specified `--target-lsn` and
`--no-get-wal` in `barman recover` command, Barman was copying all WAL
files from Barman to Postgres.

As we are able to infere the required WAL files based on the specified
`--target-lsn`, this commit changes Barman so it copies only the
required WAL files instead of them all.

References: BAR-190.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>