Add unit test for `get_required_xlog_files` with `target_xid` set
This commit adds a unit test for the `get_required_xlog_files` method
when the user specifies a `--target-xid` in `barman recover`.

In that case, the expected behavior is retrieve all WALs from the
archive up to the recovery timeline. There is no reliable or well
performant way of defining the required WAL files when dealing with
`recovery_target_xid` on Barman.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>