Add tests to the delete children backups implementation
Adding tests to test the logic of deleting children backups
when a parent is deleted. A new test was added to ensure
that the children are indeed being deleted along with the parent.
A current test was modified to make sure that, once a child has
been deleted, its reference is completely removed from its parent.

References: BAR-181

Signed-off-by: Gustavo William <gustavowilliam0980@gmail.com>