Update `walk_to_root` to optionally also yield the provided incremental
With this update, the `walk_to_root` method receives a new
parameter `include_self` that defaults to False and also yields
the provided incremental backup in the chain of backups when set
to True, which is required by `get_backup_chain`. I will still
separate the test cases for `walk_to_root` in `test_infofile.py`
in a different commit.

References: BAR-167

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>