Require cramjam lib to be >= 2.7.0 for python >= 3.6
From version 0.7.0 the library `python-snappy` has been
re-written to use internally uses the `cramjam` library.

This new version of `python-snappy` doesn't work well with versions
of `cramjam` < 2.7.0 as the structure of the code is different
than the one used in the new version.

Also for python < 3.7 new `cramjam` libs ( >= 2.7.0) are not available.

So we pinned `python-snappy` to 0.6.1 for python < 3.7, as this old version
is compatible with the `cramjam` library with versions < 2.7.0
and set to use the newer versions of the `python-snappy` and
`cramjam` libs  for python >= 3.7 .

References: bar-156

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>