Check for checksums consistency during backup and recovery
This commit introduces a new method `is_checksum_consistent` in the
`LocalBackupInfo` class which is used to check for inconsistent checksum
configurations between backups of the same chain. Having an inconsistent
chain possibly means that the user changed the value of `data_checksums`
between backups, which can lead to potential issues on the recovery phase.
The checks introduced here are currently being executed at the end of
backup and recovery executions and only issue a warning to the user,
without interrupting any process.

References: BAR-210

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>