Add information to `list-backup` about full and incremental backups
Add a new label to each backup in the list-backup output that
references the backup type considering:
  - R: for rsync backup
  - F: for full pg_basebackup
  - I: for incremental pg_basebackup
If a backup doesn't have a parent backup nor the backup method set to
rsync, it's labeled as a full backup.

This also adds a new field "backup_type" to the JSON output, containing the
backup type label.

References: BAR-207

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>