Output info from the check-backup command
The `check-backup` command can mark a backup as failed if it
notices that some of its required WAL files are missing. This works
correctly, however no output is provided, so the only way to check
why the backup has been marked as failed is by executing a
`show-backup` on the backup. This commit fixes this issue by outputting
the reason as an error message in the CLI.

References: BAR-213

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>