Add a keepalive-interval option on server config and backup command
This commit adds a `keepalive-interval` CLI argument on the `backup`
command as well as an equivalent `keepalive_interval` option on the server
configuration parameters. This new option accepts a positive integer
and will be used on an incoming keep-alive feature, which will allow
Barman to maintain a Postgres connection alive during Rsync backups.

References: BAR-192

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>