Implement a keep-alive mechanism on Rsync backups
This commit implements a keep-alive mechanism on Rsync backups. This
is necessary to guarantee that `pg_backup_start()` and `pg_backup_stop()`
are executed on the same Postgres session during an Rsync backup. This
is implemented via a new `PostgresKeepAlive` class which can be used as
context manager and will spawn a child thread to continuously send heartbeat
queries on the current connection at a specified interval during its living context.

This issue has been reported by some users and is most likely due to a
firewall rule that kills connections that have been idle for a specified
amount of time. In such cases, the stop command ends up being executed in a different
connection session than the start command and Postgres gives an error, thus the
backup in Barman is considered FAILED since the full process has not been
completed successfully and Barman is not able to guarantee its consistency.

References: BAR-192

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>