Fix next backup info of rsync incremental backups upon deletion of full
Add new method `_set_backup_sizes` that recalculates and updates the disk usage
and deduplication information of the next rsync backup after the deletion of a
backup. This ensures accurate backup information for the next rsync backup in cases
where deduplication was applied with the "reuse_backup = link" option, since that
might become outdated after the deletion. The method also includes the option
to fsync the given backup to disk, which is used by `backup_fsync_and_set_sizes`
after this update.

References: BAR-217

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>