Fix receive-wal command exiting with status 0 on errors
Fixes #539 where some errors of the `receive-wal` command were exiting with
a status of zero, indicating success. The referenced issue reports an error when using
the `--reset` option, but the same also occurs with another error when
you execute it without the flag and a receive-wal process is already
running. This commit fixes both.

References: BAR-230

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>