Fix python version in barman diagnose output
This commit fixes an issue with the `barman diagnose` command, where it
was actually getting the Python version from the first Python interpreter
found in the `PATH` environment variable instead of the one the Barman is
actually using to run. This could lead to wrong outputs if the
one found in `PATH` is different from the one that Barman is using.
This was fixed by getting the version from the `sys` module instead of
the `python --version` command which was previously.

References: BAR-232

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>