Add checks for incremental backups and compression
Backups taken with pg_basebackup can get compression. On the other
hand, incremental backups with barman cannot leverage that feature
yet because `barman recover` is not prepared to combine compressed
backups, so barman needs to prevent users from taking backups with
the `backup_compression` option set in the configuration file. For
the same reason, Barman also needs to prevent users from taking
incremental backups on top of backups that were compressed when
taken.

This commit is adding both validation checks to the
`_validate_incremental_backup_configs` method in the BackupManager
and a unit test for the use cases.

References: BAR-253

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>