Add and release keep annotation when recovering backups
An issue in github (GH-200) was reported by a user
stating that there was a failure during a recover operation due
to the fact that the backup was deleted because it entered in the
recovery window of the retention policy rule.

This commit is adding a keep annotation before the recover
operation starts and releasing the keep annotation after it ends.
This will prevent barman from deleting any backup that is being
recovered.

References: BAR-222

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>