Document changes in `barman recover` with incremental backups
This commit is adding an explanation on how the recover
from an incremental backup works using pg_combinebackup
and a later copy/move dependending on the final destination
(remote or local). Also explains about the use of the
local_staging_path for this operation.

A detailed NOTE was added to the manual where there is a
documentation about Backup commands and recover.

Also, the new field called local_staging_path was added to the
barman.1.d recover command documentation and a GUC file
for the field was added.

References: BAR-252

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>