Add new config option to config files and recover args
Since the introduction of incremental backups in PostgreSQL
17, we created a new recovery executor to handle the
operation. The plan was to reuse the `recovery_staging_path`
argument/configuration that was designed to recover
compressed backups PR#955. After further review, it was decided
that we would create a new argument/configuration specific to
recover from incremental backups called `local_staging_path`.

This commit is adding this new field to the configurations
and also in the cli argument options.
It is also changing the name of the method
`parse_recovery_staging_path` to `parse_staging_path` since
it is more generic and can be used for both `recovery_staging_path`
and `local_staging_path` without any confusion.

References: BAR-269

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>