Change implementation to accomodate new config option
After review of the way recover reuses `recovery_staging_path`,
a new field will be created to deal with recover and incremental
backups.

This commit is modifying the implementation of the recover cli
command to accomodate the new configuration option
`local_staging_path` to recover incremental backups.
Also, the unit test related to the staging paths was modified
according to the modifications made in the implementation.

References: BAR-269

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>