Add `deduplication_ratio` property to LocalBackupInfo
The `deduplication_ratio` is used to calculate the estimated
amount of resources saved when using "incremental" or "rsync"
backups. This was being calculated inside the output.py
for the console and again for the json output.

This commit is adding a new property to the LocalBackupInfo
because the `deduplication_ratio` is calculated using
backup_info fields and is related to the backup itself.

References: BAR-208

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>