Add extra fields to the get_backup_ext_info method
The "show-backup" console and json outputs make use of
the result of the get_backup_ext_info() call as the input.
Extra fields were added in this commit so we can display
it via console or json format. Some logic was moved to
this method so the code is not repeated for json AND console
outputs.

Note: The cloud-backup-show does not make use of the result
of a get_backup_ext_info() call as the input. It uses the
return value of the backup_info.to_dict() call as the input.
So this changes will not reflect in the output from the
cloud command.

References: BAR-208

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>