Change docs related to `show-backup` command new output
The docs were changed in the parts where show-backup
related information is displayed.

In the manual doc, the part where there is a mention on
show-backup command was not changed because there is no
details on the output result. A small change was made at
the "Backup metadata for snapshot backups".

A bigger change was made in the man1 where the
`show-backup` command is explained and detailed.

References: BAR-275

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>