Replace checks on `parent_backup_id` with property `is_incremental`
When the implementation for incremental backups started, we introduced
a `parent_backup_id` field to the BackupInfo class, and we had been
using that to distinguish between incremental and non-incremental
backups.

At some point we introduced a new property is_incremental which
abstracts that logic. However, that property started being used in the
new code after that, but was not used in the places where there were
already checks on ` parent_backup_id`.

This commit is replacing all checks on `parent_backup_id` for
the property `is_incremental`.

References: BAR-276

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>