Add `local_staging_path` to global/server configuration
Since the introduction of incremental backups in PostgreSQL
17, we created a new recovery executor to handle the
operation. The plan was to reuse the `recovery_staging_path`
argument/configuration that was designed to recover
compressed backups PR#955. After further review, it was decided
that we would create a new argument/configuration specific to
recover from incremental backups called `local_staging_path`.

This commit is:
	1. Add the field to the configurations.
	2. Change the name of the method
	`parse_recovery_staging_path` to `parse_staging_path`
	since it is more generic and can be used for both
	`recovery_staging_path` and `local_staging_path`
	without any confusion.
	3. Modify unit tests according to the new name
	for the method above and the usage of the
	new field `local_staging_path`.

References: BAR-269

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>