Add limitation details when using --no-get-wal with some target options
When recovering with --no-get-wal and with either of --target-xid,
--target-name or --target-time (the latter being a change introduced
by Barman 3.11), Barman will copy the whole WAL archive from the Barman
host to the recovery host.

That’s done that way because there currently is no reliable/performant
way of “guessing” before-hand the required WAL files to reach the
configured recovery target.

This commit is adding an important note to the manual doc explaining
that limitation.

References: BAR-278

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>