Improve barman logging capabilities
We have found that in certain corner cases the exceptions happening
during rsync backups execution could be not correctly logged and even
be hidden by subsequent exceptions happening in the `finally` statement
that takes care of finalizing the backup.

To avoid that, we added a `logger.debug` line that always
prints in debug mode the stack trace of exception happening
during the execution of rsync backups.

References: bar-277

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>