Fix failures in `barman-cloud-backup-delete` related with incrementals
Through Barman 3.11.0 we introduced native PG17+ incremental backups.
Some changes have been made to the backup metadata to support that.

As part of the changes, we introduced a few properties to the class
`LocalBackupInfo`. However, there was an oversight in some code paths
which are shared between local and cloud backups.

For example, when running `barman-cloud-backup-delete`, that command
reuses the retention policy classes. These, in their turn, attempt to
access the aforementioned properties.

This commit fixes the issue by moving the new properties to the base
`BackupInfo` class, so they are available both for local and for
cloud backups.

The properties `has_children` and `is_incremental` will always return
`False` for cloud backups. Analagously, `deduplication_ratio` and
`backup_type` are not expected to be used in any code path related
with cloud backups. With that in mind, moving these properties to the
base classe should not cause a problem.

References: BAR-284.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>