Add note in the manual about recovery near retention policy
This commit Closes#200. It is adding a note to the recovery
section in the documentation manual alerting the user about
the possibility of the backup getting obsolete while the
recovery process is ongoing and what to do to prevent that.

The team first thought about a solution that could handle this
automatically but many arguments were against this, specially
because it could introduce unwanted behaviour on recover and
the fact that `keep` feature is there to help.

References: BAR-222

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>